Splited header-transforms
